add leaderboard with name
