+++ src/index.js
@@ -8,9 +8,8 @@ import TitleScene from './scenes/TitleScene';
 import OptionsScene from './scenes/OptionsScene';
 import CreditsScene from './scenes/CreditsScene';
 import GameOver from './scenes/GameOver';
-import LeaderBoardScene from './scenes/leaderboard/LeaderBoard';
-import HighScore from './scenes/leaderboard/HighScore';
-import Starfield from './scenes/leaderboard/Starfield';
+import Starfield from './scenes/Starfield';
+import LeaderBoard from './scenes/LeaderBoard';
 
 class Game extends Phaser.Game {
   constructor() {
@@ -24,9 +23,8 @@ class Game extends Phaser.Game {
     this.scene.add('Credits', CreditsScene);
     this.scene.add('GameScene', GameScene);
     this.scene.add('GameOver', GameOver);
-    this.scene.add('LeaderBoard', LeaderBoardScene);
+    this.scene.add('LeaderBoard', LeaderBoard);
     this.scene.add('Starfield', Starfield);
-    this.scene.add('HighScore', HighScore);
   }
 }
