+++ src/scenes/PreloaderScene.js
@@ -86,7 +86,7 @@ export default class PreloaderScene extends Phaser.Scene {
   }
 
   ready() {
-    this.scene.start('GameScene');
+    this.scene.start('LeaderBoard');
     this.readyCount += 1;
     if (this.readyCount === 2) {
       this.scene.start('Title');
